moved search to index
